refactored fs and memory implementations out of giru.core
