+++ giru/core/commands/__init__.py
@@ -1,5 +1,4 @@
 import datetime
-import pickle
 import random
 from functools import lru_cache, reduce
 from typing import List
@@ -13,11 +12,10 @@ from telegram.message import Message
 from telegram.parsemode import ParseMode
 from telegram.bot import Bot
 
-from giru.core.scorekeeping import FsScoreKeeper
-from giru.data import julien, days, mepajeo
-from giru import data
+from giru.core.commands.data import julien, days, mepajeo
+from giru.core.ports import BaseScoreKeeper, BaseReplyStorageProvider
 from giru.helpers.movies import Movie
-from giru.settings import SPOTIPY_CLIENT_ID, SPOTIPY_CLIENT_SECRET, SCORES_FILE_PATH
+from giru.settings import SPOTIPY_CLIENT_ID, SPOTIPY_CLIENT_SECRET
 
 client_credentials = None
 
@@ -35,7 +33,7 @@ def Caps(bot, update, args):
     bot.sendMessage(chat_id=update.message.chat_id, text=text + "!")
 
 
-def create_get_saved_messages_callback(storage_provider):
+def create_get_saved_messages_callback(storage_provider: BaseReplyStorageProvider):
     def get_saved_messages_callback(bot, update):
         replies = storage_provider.get_all_replies()
 
@@ -44,7 +42,7 @@ def create_get_saved_messages_callback(storage_provider):
                 message.text, message.from_user.first_name, message.from_user.id
             )
 
-        formatted_replies = list(map(format_saved_message, replies))
+        formatted_replies = map(format_saved_message, replies)
 
         text = (
             "\n".join(formatted_replies)
@@ -170,29 +168,33 @@ def Cartelera(bot, update):
     )
 
 
-def Scores(bot, update):
-    """ Gets a list with the points scored by person. """
-    k = FsScoreKeeper(SCORES_FILE_PATH)
-    _scores = k.list_scores(update.message.chat_id)
-    if len(_scores) == 0:
+def score_command_factory(keeper: BaseScoreKeeper):
+    def scores(bot, update):
+        """ Gets a list with the points scored by person. """
+        # k = FsScoreKeeper(SCORES_FILE_PATH)
+        _scores = keeper.list_scores(update.message.chat_id)
+        if len(_scores) == 0:
+            message = "No hay scores."
+        else:
+            message = "*Scores:*\n\n"
+            sorted_scores = sorted(_scores.items(), key=lambda x: x[1], reverse=True)
+            for k, v in sorted_scores:
+                message += "*{0}:*  {1}\n".format(k, v)
+            # If it ever wants to be divided.
+            #
+            # loved = tuple(filter(lambda x: x[1] > 0, sorted_scores))
+            # hated = tuple(filter(lambda x: x[1] < 0, sorted_scores))
+            # if loved:
+            #     message += '*Loved:*\n'
+            #     for k, v in loved:
+            #         message += '{0}: {1}\n'.format(k, v)
+            # message += '\n'
+            # if hated:
+            #     message += '*Hated:*\n'
+            #     for k, v in hated:
+            #         message += '{0}: {1}\n'.format(k, v)
+        bot.sendMessage(
+            chat_id=update.message.chat_id, text=message, parse_mode="Markdown"
+        )
 
-        message = "No hay scores."
-    else:
-        message = "*Scores:*\n\n"
-        sorted_scores = sorted(_scores.items(), key=lambda x: x[1], reverse=True)
-        for k, v in sorted_scores:
-            message += "*{0}:*  {1}\n".format(k, v)
-        # If it ever wants to be divided.
-        #
-        # loved = tuple(filter(lambda x: x[1] > 0, sorted_scores))
-        # hated = tuple(filter(lambda x: x[1] < 0, sorted_scores))
-        # if loved:
-        #     message += '*Loved:*\n'
-        #     for k, v in loved:
-        #         message += '{0}: {1}\n'.format(k, v)
-        # message += '\n'
-        # if hated:
-        #     message += '*Hated:*\n'
-        #     for k, v in hated:
-        #         message += '{0}: {1}\n'.format(k, v)
-    bot.sendMessage(chat_id=update.message.chat_id, text=message, parse_mode="Markdown")
+    return scores
