+++ giru/core/commands/data.py
@@ -1,5 +1,3 @@
-""" Using this file to save data. """
-
 julien = [
     "https://media.wmagazine.com/photos/59f1031ad473f932b6e964c5/4:3/w_1536/JULIEN_WMAG_37342.jpg",
     "https://f4.bcbits.com/img/a4215132093_10.jpg",
@@ -26,22 +24,6 @@ julien = [
     "http://thebaybridged.com/wp-content/uploads/2017/02/Julien-Baker-at-GAMH-for-Noise-Pop-25-by-Ian-Young-02.jpg",
     "http://78.media.tumblr.com/119f3138c6bfe5e79e294eb287d4d0d3/tumblr_obokdgCXcP1r7o2e7o5_1280.jpg",
 ]
-
-ayuda = {
-    "vociao": {
-        "Devuelve el mensaje en mayusculas.": "/vociao [mensaje]",
-    },
-    "saved": {"Muestra los mensajes guardados.": "/saved"},
-    "spotify": {
-        "Trae el link de spotify de la cancion indicada y un preview de 30 segundos (si existe).": "/spotify [cancion]"
-    },
-    "ayuda": {"Muestra los comandos y mensajes que se pueden utilizar.": "/ayuda"},
-    "padondehoy": {"Te dice lo que vamos a hacer hoy.": "/padondehoy"},
-    "cartelera": {
-        "Trae una lista de las películas que están en los cines actualmente.": "/cartelera"
-    },
-}
-
 days = [
     # Lunes
     ["Es lunes mi loco, vete pa' tu casa.", "Duermete manin"],
@@ -58,146 +40,6 @@ days = [
     # Domingo
     ["Hoy hay iglesia, maldito impio.", "IGLESIIIIAAA"],
 ]
-
-replies = [
-    "No sé que dijiste, pero maricons ute.",
-    "Todo el mundo aquí sabe quien es el mamaguebo.",
-    "Soy un maldito bot, no sé que dices!",
-    "DAME BANDA!!",
-    "No me de reply maricons",
-    "No sé que dices",
-    "?",
-    "Qué?",
-    "Ok",
-]
-
-mmg = [
-    "mmg",
-    "mamaguevo",
-    "mamabuego",
-    "mamaguebo",
-    "mamagrano",
-    "MMG!",
-    "MMG UTE!",
-    "M M G",
-    "MMÑ",
-    "mamañema",
-    "MAMAÑEMA!",
-    "Un buen mamaguebo!",
-    "MA MA GUE BA SO",
-    "PERO MAMAGRANO!",
-    "MAMAAAAAGGGUUUUUUUEEEBO!",
-    "Mi loco no e por na pero que mmg",
-    "Diablo si que mmg",
-    "El mmg e ute en verdad",
-    "mmg pero por que",
-    "Pa mi ute lo que e un buen mamaguebo",
-    "Nunca habia conocido a alguien asi, que maldito mamaguebo",
-    "Pero que mmg",
-    "m a m a g u e b o",
-    "m a m a g r a n o",
-    "m a m a ñ e m a",
-    "Mejor me pego un tiro antes de criar tal mamaguebo",
-    "TU MALDITA MADRE MAMAGUEBO TO!",
-    "TU MALDITA MADRE",
-    "MAMAGRANO TO!",
-    "TU MALDITA MADRE MAMAÑEMA TO!",
-    "TU LO QUE ERE UN BUEN MAMAGUEBO !",
-    "TU LO QUE ERE UN BUEN MAMAGRANO!",
-    "TU LO QUE ERE UN BUEN MAMAÑEMA!",
-    "TMM MMG",
-    "Pero por favor... buen mamaguebo!",
-    "Digame de que vale ser tan mamaguebo?",
-    "Hagame el favor y ahogese con un ripio",
-    "Hay que darle una medalla por mamagrano",
-    "Mamaguebo del año",
-    "Quuuuueeeeee maaaaamaaaaaguuuuuuuueeeeebbboo",
-    "Me parece que usted es un buen mamaguebo",
-    "De vez en cuando yo creo que es medio mamaguebo si",
-    "COÑAZO CHUPESE UN RIPIO",
-    "COÑAZO MAMAGRANO",
-    "COÑAZO MAMAGUEBO",
-    "COÑAZO MAMAÑEMA",
-    "No mejor callese y mamese un guebo",
-    "Callese mamaguebo!",
-    "QUE SE CALLE MAMAGUEBO!",
-    "Una galleta por mamaguebo",
-    "Mi loco, yo voy a proceder a meterle un deo por el culo y meneaselo de lao a lao por ta' de mamaguebo",
-    "TU == MMG",
-    "Eso no e na ser mmg",
-    "Tan mamaguebo",
-    "Yo te digo, que mmg",
-    "Mamaguebo somos todos, calmese",
-    "Como se puede existir siendo tan mamaguebo",
-    "mmg... que vaina",
-    "Mire pedazo de mierda, ute no e ma que un buen mmg",
-    "Ni voy a deci na. que mamaguebo",
-    "En mi vida mis ojos habian presenciado tal nivel de mamañemismo",
-    "La vaina e que ute e un buen mmg",
-    "Cuidao que la mamañemeria se pega",
-    "Na' man, ute e el mmg. aceptelo",
-    "Por si se te olvida: TU LO QUE ERE UN BUEN MAMAGRANO!",
-    "Jesusantisima que mmg",
-    "Jesusantisima que mmñ",
-    "EL DIAAAAABLO! QUE MAMAGUEBO!",
-    "maamaguebo maamaguebo maamaguebo",
-    "maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo",
-    "maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo maamaguebo",
-    "si... no... mmg",
-    "Yo no se mmg",
-    "coño en serio mmg",
-    "Que desafine mmg",
-    "Que hacemos contigo mmg?",
-    "Toy harto de decirlo pero que maldito mamaguebo",
-    "mmg cual e la vaina?",
-    "si si si... mamañema",
-    "Ey! la mala palabra, tu maldita madre",
-    "Deje su mamañemeria por favor",
-    "No se puede con eto mamaguebos",
-    "Dejese de vaina, mamaguebo ute",
-    "Eta vaina ta llena de mamaguebos",
-    "LIBERTÉ! ÉGALITÉ! MAMAGUEBÉ!",
-    "niggas be mamaguebos",
-    "Si Dios supiera que de la descendencia del hombre iba a surgir tal gran mamaguebo, hubiese abortado a Adam en la primera semana.",
-    "Coño deberia aparecer un Hitler que se lamba a to eto mamaguebos",
-    "Hey tranquilo amigo, mas chilleo menos mamaguebismo",
-    "These violent delight have violent mamaguebos",
-    "Ma' mamaguebo que to' los Lannister juntos",
-    "Excuseme la palabra, pero que buen mamaguebo",
-    "Aveces pienso que soy un mmg tambien, pero como ute ninguno.",
-    "Erase una vez en algun pais del mundo un mamaguebo, el mamaguebo mas grande de todos... y ese mamaguebo era usted.",
-    "Record Guinness en mamañemismo.",
-    "Dicen que uno es lo que come y usted solo come ripio amiga",
-    "Por favor, mentenga este espacio 100% MAMAGUEBO FREE. Gracias.",
-    "Ay pobres incredulos mamaguebos.",
-    "Me huele a mamaguebo por aqui.",
-    "FELICIDADES! Usted es el mamaguebo 1,000,000. Se haga ganado el premio de: 3 venezolanas de las baratas! Para recibir su premio por favor mande su informacion personal y numero de cuenta a: tmm@girubot.com",
-    "FELICIDADES! Usted es el mamaguebo 1,000,000. Se haga ganado el premio de: una maldita galleta tu maldita madre! Para recibir su premio por favor mande su informacion personal y numero de cuenta a: tmm@girubot.com",
-]
-
-cposp = [
-    "aren't we all?",
-    "indeed",
-    "un buen mmg en verdad",
-    "CERTIFIED PIECE OF SHIT BOT",
-    "it's ok fam we'll be alright",
-    "nah, you are fine",
-    "Jesus made you that way",
-    "mano ute se merece una fria",
-    "*hugs*",
-    "*Giru sits in silence as an act of empathy with your human condition*",
-    "yet a bit less shit than yesterday",
-    "no le pare",
-    "no le pare, el mundo ta' lleno de mierda",
-    "you trying your best, fam",
-    "A good day to listen to Julien Baker",
-    "just gotta make it through the day and you'll be fine",
-    "quit your jobs, run to the country, start a small life",
-    "fuck that shit man",
-    "it don't matter, we are all dead already",
-    "pls get laid soon",
-]
-
 mepajeo = [
     "suelte eso",
     "baraja",
@@ -302,13 +144,4 @@ mepajeo = [
     "aquí una idea pa' que te prendas: tu en la cola de motoconcho a las 3a.m. por villa mella",
     "aquí una idea pa' que te prendas: una mujer embarazada no se puede preñar... los panas tampoco",
     "`APROVACI0N DE PAJA.... EXITOSA`",
-]
-
-BAD_CONFIG_SECRET_MESSAGE = [
-    "DISCULPE HERMANO, TENGO PROBLEMAS PARA ENTENDERLE",
-    "ERRORERRORERROERROREERROROEEROROE",
-    "NANI?!",
-    "PERO NO PUEDE SER QUE SEAN TAN MAMAÑEMA\n**flips table**",
-    "DIABLO PERO NI PORQUE YA NO LO TIENE QUE PROGRAMAR",
-    "```\nSIGNEMAVIOLATION(69420): MMG OVERFLOW.\nCRITICAL FAILURE...\nShutting down.```",
 ]
