+++ tests/test_alcohol_related_reply.py
@@ -5,7 +5,7 @@ from tests.mocks import MockMessage, MockUser
 
 class TestAlcoholRelatedFilter(TestCase):
     def test_drunk_reply_is_sent_if_filter_is_matched(self):
-        from giru.repliers import alcohol_replier
+        from giru.core.repliers import alcohol_replier
 
         matcher = alcohol_replier
         user = MockUser(first_name="Giru", is_bot=True)
