--- tests/test_get_saved_messages_command.py
@@ -1,7 +1,0 @@
-from unittest import TestCase
-
-
-class TestGetSavedMessagesCommand(TestCase):
-    def test_get_saved_commands_formats_text(self):
-        # TODO: implement this
-        pass
