+++ tests/test_scores_command.py
@@ -1,6 +1,10 @@
 from unittest import TestCase
 from unittest.mock import patch, MagicMock, mock_open
 
+from giru.adapters_memory import InMemoryScoreKeeper
+from giru.core.commands import score_command_factory
+
+from giru.core.repliers import record_points_factory
 from tests.mixins import CommandTestCaseMixin
 from tests.mocks import MockChat, MockMessage
 
@@ -10,47 +14,42 @@ load_score_keeper = MagicMock()
 
 
 class ScoresCommandTestCase(CommandTestCaseMixin, TestCase):
-    @patch("pickle.load", load_empty_scores_file)
+    def setUp(self):
+        super(ScoresCommandTestCase, self).setUp()
+        self.keeper = InMemoryScoreKeeper()
+        self.scores_command = score_command_factory(self.keeper)
+        self.record_points_replier = record_points_factory(self.keeper)
+
     def test_it_sends_a_no_score_message_if_scores_file_is_not_found(self):
         expected_chat_id = 123
         self.bot.send_message = MagicMock()
 
         update = self.create_mock_update_with_chat_id_and_message(expected_chat_id)
 
-        from giru.commands import Scores
-
-        Scores(self.bot, update)
+        self.scores_command(self.bot, update)
 
         self.bot.send_message.assert_called_with(
             chat_id=expected_chat_id, parse_mode="Markdown", text="No hay scores."
         )
 
-    @patch("pickle.load", load_scores_file)
-    @patch("builtins.open", new_callable=mock_open, read_data=b"")
-    def test_it_sends_scores(self, mock_file):
+    def test_it_sends_scores(self):
         expected_chat_id = 123
         expected_message_text = "*Scores:*\n\n" "*giru:*  1000\n"
+
+        self.keeper.scores = {"giru": 1000}
 
         self.bot.send_message = MagicMock()
 
         update = self.create_mock_update_with_chat_id_and_message(expected_chat_id)
 
-        from giru.commands import Scores
-
-        Scores(self.bot, update)
+        self.scores_command(self.bot, update)
 
         self.bot.send_message.assert_called_with(
             chat_id=expected_chat_id, parse_mode="Markdown", text=expected_message_text
         )
 
-    @patch("builtins.open", new_callable=mock_open, read_data=b"")
-    @patch(
-        "giru.core.scorekeeping.FsScoreKeeper",
-        MagicMock(return_value=load_score_keeper),
-    )
-    def test_scores_are_kept(self, mock_file):
+    def test_scores_are_kept(self):
         expected_chat_id = 123
-        expected_name = "MockUser"
 
         self.bot.send_message = MagicMock()
 
@@ -59,8 +58,6 @@ class ScoresCommandTestCase(CommandTestCaseMixin, TestCase):
             expected_chat_id, "+1", reply_to
         )
 
-        from giru.repliers import record_points
+        self.record_points_replier(self.bot, update)
 
-        record_points(self.bot, update)
-
-        load_score_keeper.add_point.assert_called_with(expected_chat_id, expected_name)
+        self.assertEqual(self.keeper.scores, {expected_chat_id: 1})
